Events: Changed the models for events
This patch modifies the model for events, now the new model
can hold more details, and can support event acknowledgement.
It also adds a audit log utility which can be used to log the
events. This stores the events in DB and also checks if
notification has to be sent for that event, if yes then it
sends a mail notification.

Change-Id: Idb74fd69c25b25f465ffca79244dc74a8e6f8412
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>