MailNotifier: Fix issue in sample mail notifier.
This patch fixes issue with sample mail notifier. Earlier
it used the old client for sending the mail(if a client was
already set). Now it initializes its own client before sending
a mail and also removed the check for mailNotification flag
as it is not needed for test mail.

Change-Id: Idfa50cf9867059cad7aec561d3586a30d6b098ff
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>