AlertClearing: Modified the utility that clears alert.
This patch modifies the utility that clears alert automatically
on corresponding recovery alert. Now if a recovery event cannot
find corresponding alert, then it sleeps for 3 minutes and checks
again this is done 5 times totalling 15 minutes.

This is done to avoid cases where clearing events are added to
DB before the corresponding alerts because they are handled by
go routines.

Change-Id: I0a016fdc2adb3c7f5465b9d1128418b5f5d215b5
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>