AlertAggregation: modified alert aggregation.
This patch modifies alert aggregation utility. All
entity models have 2 variables almWarnCount and
almCritCount instead of one almCount. With this
the alarm status for entities can be properly set.
Alert dismissal is considered for even critical and
warning events as if there is a cycle where the state
of entity changes b/w warn and crit then the count
will simply increase if we won't consider clearing.

Change-Id: I7b92c61a9491ca8c8d7b6830dc8429155bf6bb70
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>