skyring-common: Introduced timeout for ceph-installer calls.
This patch adds timeout mechanism for ceph-installer calls.
As sometimes ceph installer calls may hang, and this was
leading to inconsistant state. To avoid this, now we wait
for 15 mins for the ceph-installer call to return, if it
doesnt return we mark it as failed.
     Also the ceph-installer call during install has been
modified to be serial(earlier this was parallely done with
help of goroutines) as the ceph-installer does not support
parallel execution. Once we have the support from ceph-
isntaller this part can be reverted back.

Change-Id: Ib1ecb7ba3a636396321c059b80fad8ebe4bfee77
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>